models/imagegen/requirements.txt
kserve>=0.12.0
# torch and torchvision installed separately with CUDA index in Dockerfile
diffusers>=0.27.0
transformers>=4.40.0
pillow>=10.0.0
accelerate>=0.25.0
safetensors>=0.4.0
sentencepiece>=0.1.99
protobuf>=3.20.0
